+++ cmd/frostfs-adm/internal/commonflags/flags.go
@@ -4,4 +4,8 @@ const (
 	ConfigFlag          = "config"
 	ConfigFlagShorthand = "c"
 	ConfigFlagUsage     = "Config file"
+
+	Verbose          = "verbose"
+	VerboseShorthand = "v"
+	VerboseUsage     = "Verbose output"
 )